+++ draft-packages/button/KaizenDraft/Button/components/GenericButton.tsx
@@ -21,6 +21,9 @@ export type GenericProps = {
   fullWidth?: boolean
   disableTabFocusAndIUnderstandTheAccessibilityImplications?: boolean
   analytics?: Analytics
+  ariaControls?: string
+  ariaDescribedBy?: string
+  ariaExpanded?: boolean
   onFocus?: (e: React.FocusEvent<HTMLElement>) => void
   onBlur?: (e: React.FocusEvent<HTMLElement>) => void
 }
@@ -77,6 +80,9 @@ const renderButton: React.FunctionComponent<Props> = props => {
     onClick,
     onMouseDown,
     type,
+    ariaDescribedBy,
+    ariaExpanded,
+    ariaControls,
     disableTabFocusAndIUnderstandTheAccessibilityImplications,
     onFocus,
     onBlur,
@@ -100,6 +106,9 @@ const renderButton: React.FunctionComponent<Props> = props => {
       type={type}
       data-automation-id={props.automationId}
       title={label}
+      aria-controls={ariaControls}
+      aria-describedby={ariaDescribedBy}
+      aria-expanded={ariaExpanded}
       aria-label={label}
       tabIndex={
         disableTabFocusAndIUnderstandTheAccessibilityImplications
@@ -168,7 +177,6 @@ const buttonClass = (props: Props) => {
     [styles.reverseColorSeedling]: props.reverseColor === "seedling",
     [styles.reverseColorWisteria]: props.reverseColor === "wisteria",
     [styles.reverseColorYuzu]: props.reverseColor === "yuzu",
-    [styles.reversedIconButton]: props.iconButton && props.reversed,
   })
 }