+++ draft-packages/stories/Menu.stories.tsx
@@ -1,6 +1,9 @@
+import { Box, Paragraph } from "@kaizen/component-library"
 import * as colorTokens from "@kaizen/design-tokens/tokens/color.json"
 import { Button, IconButton } from "@kaizen/draft-button"
 const chevronDown = require("@kaizen/component-library/icons/chevron-down.icon.svg")
+  .default
+const chevronUp = require("@kaizen/component-library/icons/chevron-up.icon.svg")
   .default
 const duplicateIcon = require("@kaizen/component-library/icons/duplicate.icon.svg")
   .default
@@ -13,9 +16,15 @@ const kebabIcon = require("@kaizen/component-library/icons/kebab.icon.svg")
 const meatballsIcon = require("@kaizen/component-library/icons/meatballs.icon.svg")
   .default
 import { action } from "@storybook/addon-actions"
-import * as React from "react"
-
-import { Menu, MenuContent, MenuHeader, MenuItem, MenuSeparator } from "../menu"
+import React, { useState } from "react"
+import {
+  Menu,
+  MenuContent,
+  MenuHeader,
+  MenuItem,
+  MenuSeparator,
+  StatelessMenu,
+} from "../menu"
 
 const StoryWrapper = ({ children }) => (
   <div style={{ display: "flex", justifyContent: "center", marginTop: "1rem" }}>
@@ -77,6 +86,7 @@ export const LabelAndIcon = () => (
   <StoryWrapper>
     <Menu
       button={<Button label="Label" icon={chevronDown} iconPosition="end" />}
+      dropdownId="dropdown"
     >
       <MenuInstance />
     </Menu>
@@ -152,4 +162,99 @@ LabelAndIconBottom.story = {
   parameters: {
     viewport: { defaultViewport: "mobile1" },
   },
+}
+
+const StatelessMenuExample = props => {
+  const [isMenuVisible, setIsMenuVisible] = useState(false)
+
+  const toggleMenu = () => {
+    setIsMenuVisible(!isMenuVisible)
+  }
+
+  const hideMenu = () => {
+    setIsMenuVisible(false)
+  }
+
+  const label = `I'm ${isMenuVisible ? "open!" : "closed!"}`
+
+  return (
+    <div>
+      <Button
+        secondary={true}
+        onClick={() => toggleMenu()}
+        label="Toggle Menu"
+      />
+      <Button secondary={true} onClick={() => hideMenu()} label="Hide Menu" />
+      <Paragraph variant="body">
+        I'm {isMenuVisible ? "open!" : "closed!"}
+      </Paragraph>
+      <StatelessMenu
+        isMenuVisible={isMenuVisible}
+        toggleMenuDropdown={toggleMenu}
+        hideMenuDropdown={hideMenu}
+        renderButton={buttonProps => (
+          <Button
+            label="Label"
+            icon={isMenuVisible ? chevronUp : chevronDown}
+            iconPosition="end"
+            {...buttonProps}
+          />
+        )}
+      >
+        <div onClick={e => e.stopPropagation()}>
+          <MenuInstance />
+        </div>
+      </StatelessMenu>
+    </div>
+  )
+}
+
+export const DefaultStatelessMenu = () => {
+  return (
+    <StoryWrapper>
+      <Box p={1}>
+        <StatelessMenuExample />
+      </Box>
+      <Box p={1}>
+        <Paragraph variant="body">
+          Use the StatelessMenu component if you need to{" "}
+          <a href="https://reactjs.org/docs/lifting-state-up.html">
+            lift state
+          </a>{" "}
+          from the Menu component. This gives the flexibility to be able to
+          control the state of the dropdown however you like and respond to
+          state changes, but it requires more work to configure. It can be used
+          instead of `Menu` if this level of flexibility is required. This
+          component is used in the FilterDrawer component. View the source code{" "}
+          <a href="https://github.com/cultureamp/kaizen-design-system/blob/master/draft-packages/stories/Menu.stories.tsx">
+            here
+          </a>
+          .
+        </Paragraph>
+      </Box>
+    </StoryWrapper>
+  )
+}
+
+DefaultStatelessMenu.story = {
+  name: "StatelessMenu (example usage)",
+}
+
+export const DropdownWidthContain = () => (
+  <StoryWrapper>
+    <Menu
+      button={<Button label="Label" icon={chevronDown} iconPosition="end" />}
+      dropdownWidth="contain"
+    >
+      <MenuContent>
+        <div style={{ width: "500px" }}>
+          The dropdown is as wide as this 500px div
+        </div>
+      </MenuContent>
+    </Menu>
+  </StoryWrapper>
+)
+
+DropdownWidthContain.story = {
+  name: 'Label and Icon (dropdownWidth="contain")',
 }