+++ draft-packages/radio-group/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.32](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-radio-group@1.3.31...@kaizen/draft-radio-group@1.3.32) (2020-07-06)
+
+**Note:** Version bump only for package @kaizen/draft-radio-group
+
+
+
+
+
 ## [1.3.31](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-radio-group@1.3.30...@kaizen/draft-radio-group@1.3.31) (2020-07-01)
 
 **Note:** Version bump only for package @kaizen/draft-radio-group