+++ draft-packages/vertical-progress-step/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-vertical-progress-step@1.4.16...@kaizen/draft-vertical-progress-step@1.4.17) (2020-07-06)
+
+**Note:** Version bump only for package @kaizen/draft-vertical-progress-step
+
+
+
+
+
 ## [1.4.16](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-vertical-progress-step@1.4.15...@kaizen/draft-vertical-progress-step@1.4.16) (2020-07-01)
 
 **Note:** Version bump only for package @kaizen/draft-vertical-progress-step